디렉토리 통합
