+++ week1/김누리/baekjoon/스택_2.java
@@ -1,4 +1,4 @@
-package Week1.김누리.baekjoon;
+package week1.김누리.baekjoon;
 
 /**
  * PackageName : Week1.김누리.programmers
