+++ week1/김누리/programmers/x만큼_간격이_있는_n개의_숫자.java
@@ -1,4 +1,4 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
 import java.util.stream.IntStream;
