+++ week1/김누리/programmers/나머지가_1이_되는_수_찾기.java
@@ -1,4 +1,4 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
 import java.util.stream.IntStream;
