+++ week1/김누리/programmers/두_정수_사이의_합.java
@@ -1,4 +1,4 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
 import java.util.stream.IntStream;
