+++ week1/김누리/programmers/문자열_내_p와_y의_개수.java
@@ -1,8 +1,8 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
-/**
+/*
  * PackageName : Week1.김누리.programmers
- * FileName    : 하샤드_수
+ * FileName    : 문자열_내_p와_y의_개수
  * Author      : 김누리
  * Date        : 2025-04-25
  * Description :
@@ -11,24 +11,26 @@ package Week1.김누리.programmers;
  * ---------------------------------------------------------------------------------------------------------------------
  * 2025. 04. 25.     김누리               Initial creation
  */
-public class 하샤드_수 {
-	public boolean solution(int x) {
+public class 문자열_내_p와_y의_개수 {
+	public boolean solution(String s) {
 		boolean answer = true;
-		int harshad = 0;
-		int tmp = x;
+		int pCnt = 0;
+		int yCnt = 0;
+		String[] tmp = s.split("");
 
-		while (tmp > 0) {
-			harshad += tmp % 10; // 자리수 더하기
-			tmp /= 10;
+		for(String s2 : tmp) {
+			if(s2.equalsIgnoreCase("p")) pCnt++;
+			if(s2.equalsIgnoreCase("y")) yCnt++;
 		}
 
-		answer = x%harshad == 0 ? true : false;
+		if(pCnt != yCnt) answer = false;
 
 		return answer;
 	}
 
-	public void call_solution(int x) {
-		boolean answer = solution(x);
+	public void call_solution(String s) {
+		boolean answer = solution(s);
+
 		System.out.println(answer);
 	}
 }
