+++ week1/김누리/programmers/자릿수_더하기.java
@@ -1,4 +1,4 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
 import java.util.stream.IntStream;
