+++ week1/김누리/programmers/짝수와_홀수.java
@@ -1,8 +1,11 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
+
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 /*
  * PackageName : Week1.김누리.programmers
- * FileName    : 문자열_내_p와_y의_개수
+ * FileName    : 짝수와_홀수
  * Author      : 김누리
  * Date        : 2025-04-25
  * Description :
@@ -11,25 +14,20 @@ package Week1.김누리.programmers;
  * ---------------------------------------------------------------------------------------------------------------------
  * 2025. 04. 25.     김누리               Initial creation
  */
-public class 문자열_내_p와_y의_개수 {
-	public boolean solution(String s) {
-		boolean answer = true;
-		int pCnt = 0;
-		int yCnt = 0;
-		String[] tmp = s.split("");
+public class 짝수와_홀수 {
+	public String solution(int num) {
+		//  Stream API 이용 버전
+/*		return Stream.of(num)
+			.map(n -> num%2 == 0 ? "Even": "Odd")
+			.findFirst()
+			.get();*/
 
-		for(String s2 : tmp) {
-			if(s2.equalsIgnoreCase("p")) pCnt++;
-			if(s2.equalsIgnoreCase("y")) yCnt++;
-		}
-
-		if(pCnt != yCnt) answer = false;
-
-		return answer;
+		//  Stream API 이용하지 않은 버전
+		return num%2 == 0 ? "Even" : "Odd";
 	}
 
-	public void call_solution(String s) {
-		boolean answer = solution(s);
+	public void call_solution(int num) {
+		String answer = solution(num);
 
 		System.out.println(answer);
 	}
