+++ week1/김누리/programmers/푸드_파이트_대회.java
@@ -1,4 +1,4 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
 /**
  * PackageName : Week1.김누리.programmers
