+++ week1/김누리/programmers/하샤드_수.java
@@ -1,11 +1,8 @@
-package Week1.김누리.programmers;
+package week1.김누리.programmers;
 
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
-
-/*
+/**
  * PackageName : Week1.김누리.programmers
- * FileName    : 짝수와_홀수
+ * FileName    : 하샤드_수
  * Author      : 김누리
  * Date        : 2025-04-25
  * Description :
@@ -14,21 +11,24 @@ import java.util.stream.Stream;
  * ---------------------------------------------------------------------------------------------------------------------
  * 2025. 04. 25.     김누리               Initial creation
  */
-public class 짝수와_홀수 {
-	public String solution(int num) {
-		//  Stream API 이용 버전
-/*		return Stream.of(num)
-			.map(n -> num%2 == 0 ? "Even": "Odd")
-			.findFirst()
-			.get();*/
+public class 하샤드_수 {
+	public boolean solution(int x) {
+		boolean answer = true;
+		int harshad = 0;
+		int tmp = x;
 
-		//  Stream API 이용하지 않은 버전
-		return num%2 == 0 ? "Even" : "Odd";
+		while (tmp > 0) {
+			harshad += tmp % 10; // 자리수 더하기
+			tmp /= 10;
+		}
+
+		answer = x%harshad == 0 ? true : false;
+
+		return answer;
 	}
 
-	public void call_solution(int num) {
-		String answer = solution(num);
-
+	public void call_solution(int x) {
+		boolean answer = solution(x);
 		System.out.println(answer);
 	}
 }
